Created working blinds kind of
so added blinds model, it will open and close when you 'press'' the open or close buttons in the editor window. so it should work when in actual vr. however the problem im having is when i add another blind prefab, and then press the open and close buttons, it doesnt work the other blinds ive duplicated, only works for the original one.